% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.GmultiREEMtree.R
\name{fitted.GmultiREEMtree}
\alias{fitted.GmultiREEMtree}
\title{Return the fitted values of a GmultiREEMtree object}
\usage{
\method{fitted}{GmultiREEMtree}(object, level = -1, ...)
}
\arguments{
\item{object}{A GmultiREEMtree object.}

\item{level}{An optional integer giving the level of grouping to be used in
extracting the fitted values from object. Defaults to be the innermost level (level=-1).}

\item{...}{Won't be used.}
}
\value{
A matrix of fitted values.
}
\description{
Return the fitted values of a GmultiREEMtree object.
}
