% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.multiREEMtree.R
\name{fitted.multiREEMtree}
\alias{fitted.multiREEMtree}
\title{Return the fitted values of a multiREEMtree object}
\usage{
\method{fitted}{multiREEMtree}(object, level = -1, ...)
}
\arguments{
\item{object}{A multiREEMtree object.}

\item{level}{An optional integer giving the level of grouping to be used in
extracting the fitted values from object. Defaults to be the innermost level (level=-1).
If the lme.method of the object is "MCMCglmm", fitted levels can only be 0 (for fixed) or 1
(for fixed + random).}

\item{...}{Won't be used.}
}
\value{
A matrix of fitted values.
}
\description{
Return the fitted values of a multiREEMtree object.
}
\examples{
require(multiREEMtree)
# simulate multivariate grouped data
X1 <- rnorm(100, 0, 1)
X2 <- rnorm(100, 0, 1)
X3 <- rnorm(100, 0, 1)
group <- rep(1:10, 10)
simb <- MASS::mvrnorm(10, rep(0, 2), diag(2))
sime <- MASS::mvrnorm(100, rep(0, 2), diag(2))
Y1 <- 10 * (X1>0) + 20 * (X1<0 & X2 >0) + simb[group, 1] + sime[ ,1]
Y2 <- 30 * (X1>0) + 50 * (X1<0 & X2 >0)+ simb[group, 2] + sime[ ,2]
data <- data.frame(cbind(X1, X2, X3, Y1, Y2, group))
f <- Y1 + Y2 ~ X1 + X2 + X3
r <-  ~ 1 | group
result <- multiREEMtree(formula=f, data=data[1:90, ], random=r)
fit <- fitted(result)
}
