% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.multitree.R
\name{fitted.multitree}
\alias{fitted.multitree}
\title{Return the fitted values of a multitree object}
\usage{
\method{fitted}{multitree}(object, ...)
}
\arguments{
\item{object}{A multitree object.}

\item{...}{Won't be used.}
}
\value{
A matrix of fitted values.
}
\description{
Return the fitted values of a multitree object.
}
