% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmultiREEMtree.R
\name{GmultiREEMtree}
\alias{GmultiREEMtree}
\title{Generalized Multivariate Random Effects EM Tree for longitudinal and clustered data}
\usage{
GmultiREEMtree(
  formula,
  data,
  random,
  subset = NULL,
  link = "logit",
  InitialEta = "default",
  InitialRandomEffects = 0,
  ErrorTolerance = 0.001,
  MaxIterations = 1000,
  tree.xv = "1se",
  tree.xval = 10,
  tree.control = rpart.control(cp = 0.01, maxsurrogate = 5, usesurrogate = 2),
  autoCorrelation = NULL,
  lme.method = "REML",
  lme.na.action = na.fail,
  lme.control = lmeControl(opt = "optim")
)
}
\arguments{
\item{formula}{A formula of the form \code{y1 + ... + ym ~ x1 + ... + xn}, where \code{y1,...,ym} are response
variables and \code{x1,...,xn} are predictors.}

\item{data}{A data frame that contains all of the variables named in the formula.}

\item{random}{A one-sided formula of the form \code{~ z1 + ... + zn | g1/.../gm} with \code{z1 + ... + zn} specifying the model
for the random effects and \code{g1/.../gm} the grouping structure (\code{m} may be equal to 1, in which case no / is required).
\code{z1,...,zn} and \code{g1,...,gm} should be contained in data. Use \code{~ 1 | g1/.../gm} to model only a random intercept for each group
and use \code{~ 0 + z1 + ... + zn | g1/.../gm} to remove the random intercept.}

\item{subset}{A logical vector showing the subset of data to use for fitting. Defaults to be a vector of TRUEs.}

\item{link}{A link function connnecting the expectation of the response variable and a linear predictor. Can either be \code{"logit"},
\code{"log"} or a self-defined list containing three elements: \code{"h"} -- a univariate function that satisfies E(Y|X, b)=h(eta);
\code{"dh"} -- the first-order derivative of h; \code{"h_inv"} -- the inverse of h.}

\item{InitialEta}{A vector of initial estimations for eta. Defaults to be "default," which sets eta=0.75 if y=1 and eta=0.25 if y=0.}

\item{InitialRandomEffects}{A vector of initial values for random effects, with default value 0.}

\item{ErrorTolerance}{The error tolerance for the stopping criterion of the algorithm. When the
increase of loglikelihood for "nlme" is less than \code{ErrorTolerance},
the algorithm will stop. Defaults to be 0.001.}

\item{MaxIterations}{The largest number of iterations. Defaults to be 1000.}

\item{tree.xv}{Selection of tree by cross-validation: \code{"1se"} (default) - gives the simplest tree that has cross-validation error within
one SE of the minimum value using k-fold cross-validation, with k given by tree.xval; \code{"min"} - the tree with minimum error
of k-fold cross-validation, with k given by tree.xval; \code{"none"} - no cross-validation.}

\item{tree.xval}{Number of cross-validation folds or vector defining cross-validation groups. Defaults to be 10.
Could also be \code{"LOOCV"}, which will use leave-one-out cross-validation.}

\item{tree.control}{A list of control values used to control the \link[mvpart]{rpart}; See \link[mvpart]{rpart.control}
for details. Defaults to be \code{rpart.control(cp=0.01, maxsurrogate=5, usesurrogate=2)}.}

\item{autoCorrelation}{An optional string specifying the within-object correlation structure. Should be either NULL (default) or the name
of one of the available corStruct classes in \link[nlme]{corClasses} (e.g., "corAR1"). This auto-correlation structure will be applied to all of
the objects and response variables.}

\item{lme.method}{The method used in \link[nlme]{lme}. Defaults to be \code{"REML"}. Can also be \code{"ML"}.}

\item{lme.na.action}{A function that indicates what should happen when the data contain NAs. The default action (na.fail) causes lme
to print an error message and terminate if there are any incomplete observations. See \link[nlme]{lme}
for other options.}

\item{lme.control}{A list of control values used to control the \link[nlme]{lme} algorithm, with default opt="optim".
See \link[nlme]{lmeControl} for details.}
}
\value{
A GmultiREEMtree object, which includes
\itemize{
\item Tree - The fitted multivariate tree.
\item EffectModel - The fitted linear mixed effects model.
\item RandomEffects - A list of the random effects of each response variable and
each random level.
\item fitted - A list of the fitted values for the expectation of the responses
of, with all of the fitted levels. The j-th item is a matrix equivalent to the output of using
\code{fitted()} with level=j-1. (level=-1 represents the innermost level.)
\item ResponseVariables - The response variables \code{y1,...,ym} defined by \code{formula}.
\item RandomPredictors - The random predictors \code{z1,...,zn} defined by \code{random}.
\item GroupVariables - Group Variables \code{g1/.../gm} in \code{random}.
\item data - The data set.
\item logLik - The final loglikelihood.
\item IterationsUsed - The number of iterations used in the algorithm.
\item ErrorTolerance - The parameter \code{ErrorTolerance} in the function.
\item link - The parameter \code{link} in the function.
\item Formula - The parameter \code{formula} in the function.
\item Random - The parameter \code{random} in the function.
\item Subset - The parameter \code{Subset} in the function.
\item tree.control- \code{tree.control} in the function.
\item xv - The parameter \code{tree.xv} in the function.
\item xval - The parameter \code{tree.xval} in the function.
\item lme.method - The parameter \code{lme.method} in the function.
\item autoCorrelation - The fitted autoCorrelation structure.
\item lme.control- The parameter \code{lme.control} in the function.
\item lme.na.action - The parameter \code{lme.na.action} in the function.
\item Y_original - The original values of the response variables.
}
}
\description{
\code{GmultiREEMtree} This program implements the generalized multivariate RE-EM tree method in Jing and Simonoff (2022) for
non-Gaussian response variables. It is assumed that E(Y|X, b)=h(eta) and eta=f(X)+Zb,
where Y is the response variable, h is a given function, f(X) is a multivariate regression tree, and b is the random effect.
}
\examples{
require(multiREEMtree)
# simulate multivariate grouped data with binary response
X1 <- rnorm(1000, 0, 1)
X2 <- rnorm(1000, 0, 1)
X3 <- rnorm(1000, 0, 1)
group <- rep(1:10, 100)
simb <- MASS::mvrnorm(10, rep(0, 2), 0.01 * diag(2))
sime <- MASS::mvrnorm(1000, rep(0, 2), 0.001 * diag(2))
eta1 <- 2 * (X1>0) + (-2) * (X1<0 & X2 >0) + simb[group, 1] + sime[ ,1]
eta2 <- 2 * (X1>0) + (-2) * (X1<0 & X2 >0) + simb[group, 2] + sime[ ,2]
Y1 <- rep(NA, 1000)
Y2 <- rep(NA, 1000)
mu1 <- exp(eta1) / (exp(eta1) + 1)
mu2 <- exp(eta2) / (exp(eta2) + 1)
for (i in 1:1000) {
  Y1[i] <- rbinom(1, 1, mu1[i])
  Y2[i] <- rbinom(1, 1, mu2[i])
}
data <- data.frame(cbind(X1, X2, X3, Y1, Y2, group))
f <- Y1 + Y2 ~ X1 + X2 + X3
r <-  ~ 1 | group
result <- GmultiREEMtree(formula=f, data=data, random=r)
}
\references{
Jing, W. and Simonoff, J. S.  (2022). "A Regression Tree Method for Longitudinal and Clustered Data With Multivariate Responses." arXiv.

Sela, R. J. and Simonoff, J. S. (2012). "RE-EM Trees: A Data Mining Approach for Longitudinal and Clustered Data." \emph{Machine Learning}, \strong{86}, 169-207.
}
