% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.GmultiREEMtree.R
\name{is.GmultiREEMtree}
\alias{is.GmultiREEMtree}
\title{Check if an R object is a GmultiREEMtree object}
\usage{
is.GmultiREEMtree(object)
}
\arguments{
\item{object}{An object to be checked.}
}
\value{
A logical value
}
\description{
Check if an R object is a GmultiREEMtree object.
}
