% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.multitree.R
\name{is.multitree}
\alias{is.multitree}
\title{Check if an R object is a multitree object}
\usage{
is.multitree(object)
}
\arguments{
\item{object}{An object to be checked.}
}
\value{
A logical value
}
\description{
Check if an R object is a multitree object.
}
