% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiREEMtree.R
\name{multiREEMtree}
\alias{multiREEMtree}
\title{Multivariate Random Effects EM Tree for longitudinal and clustered data}
\usage{
multiREEMtree(
  formula,
  data,
  random,
  subset = NULL,
  stand.y = "marginal",
  InitialRandomEffects = 0,
  ErrorTolerance = 0.001,
  MaxIterations = 1000,
  tree.xv = "1se",
  tree.xval = 10,
  tree.control = rpart.control(cp = 0.01, maxsurrogate = 5, usesurrogate = 2),
  lme.algorithm = "nlme",
  autoCorrelation = NULL,
  lme.method = "REML",
  lme.na.action = na.fail,
  lme.control = lmeControl(opt = "optim")
)
}
\arguments{
\item{formula}{A formula of the form \code{y1 + ... + ym ~ x1 + ... + xn}, where \code{y1,...,ym} are response
variables and \code{x1,...,xn} are predictors.}

\item{data}{A data frame that contains all of the variables named in the formula.}

\item{random}{A one-sided formula of the form \code{~ z1 + ... + zn | g1/.../gm} with \code{z1 + ... + zn} specifying the model
for the random effects and \code{g1/.../gm} the grouping structure (\code{m} may be equal to 1, in which case no / is required).
\code{z1,...,zn} and \code{g1,...,gm} should be contained in data. Use \code{~ 1 | g1/.../gm} to model only a random intercept for each group
and use \code{~ 0 + z1 + ... + zn | g1/.../gm} to remove the random intercept. If \code{lme.algorithm="MCMCglmm"}, only \code{m}=1 is allowed.}

\item{subset}{A logical vector showing the subset of data to use for fitting. Defaults to be a vector of TRUEs.}

\item{stand.y}{An option to standardize the response variables, should be one of "marginal" (standardize each response variable by subtracting
its mean and dividing by its sample standard deviation), "covariance"(standardize the set of response variables using the vector of sample
means and sample covariance matrix, i.e., convert y to y * (cov(y)^(-1/2)), or "none" (no standardization). Defaults to be "marginal".}

\item{InitialRandomEffects}{A vector of initial values for random effects, with default value 0.}

\item{ErrorTolerance}{The error tolerance for the stopping criterion of the algorithm. When the
increase of loglikelihood for "nlme" (or the decrease of DIC for "MCMCglmm") is less than \code{ErrorTolerance},
the algorithm will stop. Defaults to be 0.001.}

\item{MaxIterations}{The largest number of iterations. Defaults to be 1000.}

\item{tree.xv}{Selection of tree by cross-validation: \code{"1se"} (default) - gives the simplest tree that has cross-validation error within
one SE of the minimum value using k-fold cross-validation, with k given by tree.xval; \code{"min"} - the tree with minimum error
of k-fold cross-validation, with k given by tree.xval; \code{"none"} - no cross-validation.}

\item{tree.xval}{Number of cross-validation folds or vector defining cross-validation groups. Defaults to be 10.
Could also be \code{"LOOCV"}, which will use leave-one-out cross-validation.}

\item{tree.control}{A list of control values used to control the \link[mvpart]{rpart}; See \link[mvpart]{rpart.control}
for details. Defaults to be \code{rpart.control(cp=0.01, maxsurrogate=5, usesurrogate=2)}.}

\item{lme.algorithm}{The algorithm used to fit linear mixed effect model. Either \code{"nlme"} or \code{"MCMCglmm"}. Defaults to be "nlme".}

\item{autoCorrelation}{An optional string specifying the within-object correlation structure. Should be either NULL (default) or the name
of one of the available corStruct classes in \link[nlme]{corClasses} (e.g., "corAR1"). This auto-correlation structure will be applied to all of
the objects and response variables. Only used when \code{lme.algorithm="nlme"}.}

\item{lme.method}{The method used in \link[nlme]{lme}. Defaults to be \code{"REML"}. Can also be \code{"ML"}. Only used when \code{lme.algorithm="nlme"}.}

\item{lme.na.action}{A function that indicates what should happen when the data contain NAs. The default action (na.fail) causes lme
to print an error message and terminate if there are any incomplete observations. Only used when \code{lme.algorithm="nlme"}. See \link[nlme]{lme}
for other options.}

\item{lme.control}{A list of control values used to control the \link[nlme]{lme} algorithm, with default opt="optim".
See \link[nlme]{lmeControl} for details. Only used when \code{lme.algorithm="nlme"}.}
}
\value{
A multiREEMtree object, which includes.
\itemize{
\item Tree - The fitted multivariate tree.
\item EffectModel - The fitted linear mixed effects model.
\item RandomEffects - A list of the random effects of each response variable and
each random level.
\item fitted - A list of the fitted values of all of the levels.
The j-th item is a matrix equivalent to the output of using
\code{fitted()} with level=j-1. (level=-1 represents the innermost level.)
\item residuals - A list of the residual values of all of the levels. The
j-th item is a matrix equivalent to the output of using
\code{residual()} with level=j-1.
\item ResponseVariables - The response variables \code{y1,...,ym} defined by \code{formula}.
\item RandomPredictors - The random predictors \code{z1,...,zn} defined by \code{random}.
\item GroupVariables - Group Variables \code{g1/.../gm} in \code{random}.
\item data - The data set.
\item LogLik - The final loglikelihood if \code{lme.algorithm="nlme"}.
\item DIC - The final value of Deviance Information Criterion if \code{lme.algorithm="MCMCglmm"}.
\item IterationsUsed - The number of iterations used in the algorithm.
\item ErrorTolerance - The parameter \code{ErrorTolerance} in the function.
\item Formula - The parameter \code{formula} in the function.
\item Random - The parameter \code{random} in the function.
\item Subset - The parameter \code{Subset} in the function.
\item tree.control- \code{tree.control} in the function.
\item xv - The parameter \code{tree.xv} in the function.
\item xval - The parameter \code{tree.xval} in the function.
\item lme.algorithm - The parameter \code{lme.algorithm} in the function.
\item lme.method - The parameter \code{lme.method} in the function.
\item autoCorrelation - The fitted autoCorrelation structure.
\item lme.control- The parameter \code{lme.control} in the function.
\item lme.na.action - The parameter \code{lme.na.action} in the function.
\item Y_original - The original values of the response variables.
\item stand.y - The parameter \code{stand.y} in the function.
}
}
\description{
\code{multiREEMtree} This program implements the multivariate RE-EM method of Jing and Simonoff (2022),
which extends the RE-EM algorithm of Sela and Simonoff (2012) to a multivariate set of response variables.
}
\examples{
require(multiREEMtree)
# simulate multivariate grouped data
X1 <- rnorm(100, 0, 1)
X2 <- rnorm(100, 0, 1)
X3 <- rnorm(100, 0, 1)
group <- rep(1:10, 10)
simb <- MASS::mvrnorm(10, rep(0, 2), diag(2))
sime <- MASS::mvrnorm(100, rep(0, 2), diag(2))
Y1 <- 10 * (X1>0) + 20 * (X1<0 & X2 >0) + simb[group, 1] + sime[ ,1]
Y2 <- 30 * (X1>0) + 50 * (X1<0 & X2 >0)+ simb[group, 2] + sime[ ,2]
data <- data.frame(cbind(X1, X2, X3, Y1, Y2, group))
f <- Y1 + Y2 ~ X1 + X2 + X3
r <-  ~ 1 | group
result <- multiREEMtree(formula=f, data=data, random=r)
}
\references{
Jing, W. and Simonoff, J. S.  (2022). "A Regression Tree Method for Longitudinal and Clustered Data With Multivariate Responses." arXiv.

Sela, R. J. and Simonoff, J. S. (2012). "RE-EM Trees: A Data Mining Approach for Longitudinal and Clustered Data." \emph{Machine Learning}, \strong{86}, 169-207.
}
