% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitree.R
\name{multitree}
\alias{multitree}
\title{Wrapper function for calling of mvpart function}
\usage{
multitree(
  formula,
  data,
  subset = NULL,
  stand.y = "marginal",
  tree.xv = "1se",
  tree.xval = 10,
  tree.control = rpart.control(cp = 0.01, maxsurrogate = 5, usesurrogate = 2)
)
}
\arguments{
\item{formula}{A formula of the form \code{y1 + ... + ym ~ x1 + ... + xn}, where \code{y1,...,ym} are response
variables and \code{x1,...,xn} are predictors.}

\item{data}{A data frame that contains all of the variables named in the formula.}

\item{subset}{A logical vector showing the subset of data to use for fitting. Defaults to be a vector of TRUEs.}

\item{stand.y}{An option to standardize the response variables, should be one of "marginal" (standardize each response variable by subtracting
its mean and dividing by its sample standard deviation), "covariance"(standardize the set of response variables using the vector of sample
means and sample covariance matrix, i.e., convert y to y * (cov(y)^(-1/2)), or "none" (no standardization). Defaults to be "marginal".}

\item{tree.xv}{Selection of tree by cross-validation: \code{"1se"} (default) - gives the simplest tree that has cross-validation error within
one SE of the minimum value using k-fold cross-validation, with k given by tree.xval; \code{"min"} - the tree with minimum error
of k-fold cross-validation, with k given by tree.xval; \code{"none"} - no cross-validation.}

\item{tree.xval}{Number of cross-validation folds or vector defining cross-validation groups. Defaults to be 10.
Could also be \code{"LOOCV"}, which will use leave-one-out cross-validation.}

\item{tree.control}{A list of control values used to control the \link[mvpart]{rpart}; See \link[mvpart]{rpart.control}
for details. Defaults to be \code{rpart.control(cp=0.01, maxsurrogate=5, usesurrogate=2)}.}
}
\value{
A multitree object, which includes.
\itemize{
\item Tree - The fitted multivariate tree.
\item ResponseVariables - The response variables \code{y1,...,ym} defined by \code{formula}.
\item Predictors - The predictors \code{x1,...,xn} defined by \code{formula}.
\item data - The data set.
\item Formula - The parameter \code{formula} in the function.
\item Subset - The parameter \code{Subset} in the function.
\item tree.control- \code{tree.control} in the function.
\item xv - The parameter \code{tree.xv} in the function.
\item xval - The parameter \code{tree.xval} in the function.
\item Y_original - The original values of the response variables.
\item stand.y - The parameter \code{stand.y} in the function.
}
}
\description{
Wrapper function for calling of mvpart function
}
