% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GmultiREEMtree.R
\name{plot.GmultiREEMtree}
\alias{plot.GmultiREEMtree}
\title{Plot the multivariate tree of a GmultiREEMtree object}
\usage{
\method{plot}{GmultiREEMtree}(x, ...)
}
\arguments{
\item{x}{A GmultiREEMtree object.}

\item{...}{Identical to the arguments in \link[mvpart]{text.rpart}}
}
\value{
A plot of a multivariate RE-EM tree.
}
\description{
Plot the multivariate tree of a GmultiREEMtree object.
}
