% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.multiREEMtree.R
\name{predict.multiREEMtree}
\alias{predict.multiREEMtree}
\title{Make predictions of a multiREEMtree object on a new dataset}
\usage{
\method{predict}{multiREEMtree}(object, newdata, level = -1, ...)
}
\arguments{
\item{object}{A multiREEMtree object.}

\item{newdata}{A data frame in which to look for variables with which to predict.
All variables used in the fixed and random effects models, as well as the grouping
factors, must be included in the data frame.}

\item{level}{An optional integer giving the level of grouping to be used
in obtaining the predictions. Defaults to be the innermost level (level=-1). If the lme.method
of the object is "MCMCglmm", predicting levels can only be 0 (for fixed) or 1 (for fixed + random).}

\item{...}{Won't be used.}
}
\value{
A matrix of predicted values.
}
\description{
Make predictions of a multiREEMtree object on a new dataset.
}
\examples{
require(multiREEMtree)
# simulate multivariate grouped data
X1 <- rnorm(100, 0, 1)
X2 <- rnorm(100, 0, 1)
X3 <- rnorm(100, 0, 1)
group <- rep(1:10, 10)
simb <- MASS::mvrnorm(10, rep(0, 2), diag(2))
sime <- MASS::mvrnorm(100, rep(0, 2), diag(2))
Y1 <- 10 * (X1>0) + 20 * (X1<0 & X2 >0) + simb[group, 1] + sime[ ,1]
Y2 <- 30 * (X1>0) + 50 * (X1<0 & X2 >0)+ simb[group, 2] + sime[ ,2]
data <- data.frame(cbind(X1, X2, X3, Y1, Y2, group))
f <- Y1 + Y2 ~ X1 + X2 + X3
r <-  ~ 1 | group
result <- multiREEMtree(formula=f, data=data[1:90, ], random=r)
pred <- predict(result, newdata=data[91:100, ], level=1)
}
