% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.multitree.R
\name{predict.multitree}
\alias{predict.multitree}
\title{Make predictions based on a multitree object on a new dataset}
\usage{
\method{predict}{multitree}(object, newdata, ...)
}
\arguments{
\item{object}{A multiree object.}

\item{newdata}{A data frame in which to look for variables with which to predict.}

\item{...}{Won't be used.}
}
\value{
A matrix of predicted values.
}
\description{
Make predictions based on a multitree object on a new dataset.
}
