% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomEffect.R
\name{RandomEffect}
\alias{RandomEffect}
\title{Return the predicted random effects of a multiREEMtree or GmultiREEMtree object}
\usage{
RandomEffect(object)
}
\arguments{
\item{object}{A multiREEMtree or GmultiREEMtree object.}
}
\value{
A list recording the predicted random effects in a multiREEMtree or GmultiREEMtree object.
}
\description{
Return the predicted random effects of a multiREEMtree or GmultiREEMtree object
}
