% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.GmultiREEMtree.R
\name{residuals.GmultiREEMtree}
\alias{residuals.GmultiREEMtree}
\title{Return the residuals of a GmultiREEMtree object}
\usage{
\method{residuals}{GmultiREEMtree}(object, level = -1, ...)
}
\arguments{
\item{object}{A GmultiREEMtree object.}

\item{level}{An optional integer giving the level of grouping to be used in
extracting the residuals from object. Defaults to be the innermost level.}

\item{...}{Won't be used.}
}
\value{
A matrix of residuals.
}
\description{
Return the residuals of a GmultiREEMtree object
}
