% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.multitree.R
\name{residuals.multitree}
\alias{residuals.multitree}
\title{Return the residuals of a multitree object}
\usage{
\method{residuals}{multitree}(object, ...)
}
\arguments{
\item{object}{A multitree object.}

\item{...}{Won't be used.}
}
\value{
A matrix of residuals.
}
\description{
Return the residuals of a multitree object
}
