% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.GmultiREEMtree.R
\name{summary.GmultiREEMtree}
\alias{summary.GmultiREEMtree}
\title{Summarize a GmultiREEMtree object}
\usage{
\method{summary}{GmultiREEMtree}(object, ...)
}
\arguments{
\item{object}{a GmultiREEMtree object.}

\item{...}{Won't be used.}
}
\description{
Summarize a GmultiREEMtree object.
}
