% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.multiREEMtree.R
\name{summary.multiREEMtree}
\alias{summary.multiREEMtree}
\title{Summarize a multiREEMtree object}
\usage{
\method{summary}{multiREEMtree}(object, ...)
}
\arguments{
\item{object}{a multiREEMtree object.}

\item{...}{Won't be used.}
}
\description{
Summarize a multiREEMtree object.
}
