% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.multitree.R
\name{summary.multitree}
\alias{summary.multitree}
\title{Summarize a multitree object}
\usage{
\method{summary}{multitree}(object, ...)
}
\arguments{
\item{object}{a multiree object.}

\item{...}{Won't be used.}
}
\description{
Summarize a multitree object.
}
