% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTerminal.R
\name{getTerminal}
\alias{getTerminal}
\title{Extract the terminal nodes (leaves) of a multiREEMtree (or multitree) object}
\usage{
getTerminal(object)
}
\arguments{
\item{object}{A multiREEMtree or multitree object.}
}
\value{
A list containing
\itemize{
\item terminal_nodes - a character vector showing all of the terminal nodes.
\item terminal_nodes_table - a table showing the number of observations in each terminal node.
\item terminal_nodes_where - a vector showing the terminal node to which each observation belongs.
\item terminal_nodes_summary - a matrix summarizing the mean and standard deviation of each
response variable in each terminal node.

}
}
\description{
Extract the terminal nodes (leaves) of a multiREEMtree (or multitree) object, and provide summary statistics for
each response variable in each node.
}
