% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomEffect.R
\name{RandomEffect}
\alias{RandomEffect}
\title{Return the random effect of a multiREEMtree object}
\usage{
RandomEffect(object)
}
\arguments{
\item{object}{A multiREEMtree object.}
}
\value{
A list recording the random effects in a multiREEMtree object.
}
\description{
Return the random effect of a multiREEMtree object
}
